// src/app/private/components/veiculo/model/veiculo.ts
export interface Veiculo {

    id: string;
    placa: string;
    renavam: string;
    chassi: string;
    tipoChassi: string;
    ano: number;
    modelo: string;
    marca: string;
    combustivel: string;
    ultimaKm: string;
    corInterna?: string;
    corExterna?: string;
    numMotorInterno?: number;
    numMotorExterno?: number;
    rastreador?: boolean;
    bloqueador?: boolean;
    dataAquisicao: string;
    condicao: string;
    valorCompra: number;
    ValorReforma?: number;
    valorMercado: number;
    nomeVendedor: string;
    teleVendedor: string;
    descricao: string;
}

export interface Veiculo2 {

    placa: string;
    renavam: string;
    chassi: string;
    tipoChassi: string;
    ano: number;
    modelo: string;
    marca: string;
    combustivel: string;
    ultimaKm: string;
    corInterna?: string;
    corExterna?: string;
    numMotorInterno?: number;
    numMotorExterno?: number;
    rastreador?: boolean;
    bloqueador?: boolean;
    dataAquisicao: string;
    condicao: string;
    valorCompra: number;
    ValorReforma?: number;
    valorMercado: number;
    nomeVendedor: string;
    teleVendedor: string;
    descricao: string;
}